+++ R/dbi-driver.R
@@ -74,7 +74,7 @@ setMethod(
 #' @export
 setMethod(
   "dbConnect", "BigQueryDriver",
-  function(drv, project, dataset, billing = project,
+  function(drv, project, dataset = NULL, billing = project,
            page_size = 1e4,
            quiet = NA,
            use_legacy_sql = FALSE,