+++ tests/testthat/test-dbi-connection.R
@@ -29,3 +29,16 @@ test_that("can roundtrip a data frame", {
   expect_equal(nrow(df), 32)
   expect_equal(ncol(df), 11)
 })
+
+test_that("dataset is optional", {
+  con <- DBI::dbConnect(bigquery(), project = bq_test_project())
+  expect_error(DBI::dbListTables(con), "`dataset`")
+
+  df <- DBI::dbReadTable(con, "publicdata.samples.natality", max_results = 10)
+  expect_equal(ncol(df), 31)
+
+  expect_error(
+    DBI::dbReadTable(con, "natality", max_results = 10),
+    "must have 2 or 3 components"
+  )
+})